Fixed warnings generated by ax_llvm when checking for libraries.
The script which checks that all libraries which are required by LLVM
can be found, caused warnings by using AC_CHECK_LIB in an unsupported
way (because the token to check for in the library was not known). Now
we check for libraries by directly trying to link against them
instead.