Add support for travis
This pull request adds support for automatically building nidhugg and
running the test on travis.org for pull requests and for commits to
the master branch of the repository.

As a first step, it adds support for testing under two relatively old
LLVM versions (3.5.0 and 3.6.2).  Newer ones (e.g., 3.8.1) can currently
not be run on Travis because of llvm-config returning some unknown flags
which g++ cannot handle.