add nidhugg support for passing arguments to the test program
This commit makes it possible to pass arguments to the test
program. The following examples illustrate how this can be done:

$ nidhugg --sc test_program.ll -- argument1 argument2

$ nidhugg --sc test_program.ll -- --argument1 --argument2

The commit also add the name of the source file (.ll file) as the
first program argument when a source file is specified. The commit
does not change the meaning of previously valid command lines in any
other way.